requirements/test.txt
Django==1.8.2
markdown2==2.3.0
Pygments==2.0.2
wheel==0.24.0
psycopg2==2.6.1

django-jenkins==0.18.1

coverage==4.3.4
coveralls==1.1
docopt==0.6.2

feedparser==5.2.1
factory_boy==2.8.1

django-crispy-forms==1.6.1

django-axes==2.3.2
pytz==2016.10
